#213 Added email validation for already existing users in add user modal
