+++ UInnovateApp/src/enums/ErrMsg.ts
@@ -1,0 +1,9 @@
+export enum ErrMsg {
+	INVALID_EMAIL = "Invalid email address",
+	EMAIL_NOT_FOUND = "Couldn't find your email address in the system",
+	USER_ALREADY_EXISTS = "User already exists",
+	WRONG_PASSWORD = "Incorrect password for the given email address",
+	INSECURE_PASSWORD = "Password doesn't meet the requirements",
+	NO_MATCH_CONFIRM_PASSWORD = "Passwords don't match",
+	MISSING_FIELD = "Missing field"
+}
